major chnages to how images are stored... still have a few kinks to work out

+++ lib/database/recipe_photo_database_manager.dart
@@ -1,3 +1,6 @@
+import 'dart:convert';
+import 'dart:core';
+
 import 'package:my_recipes/database/recipe_database.dart';
 import 'package:my_recipes/model/recipe_photo.dart';
 import 'package:sqflite/sqflite.dart';
@@ -15,9 +18,25 @@ class RecipePhotoDatabaseManager {
       return RecipePhoto(
           id: maps[i]['id'],
           recipeId: recipeId,
-          value: maps[i]['value'],
+          image:
+              maps[i]['image'] != null ? base64.decode(maps[i]['image']) : null,
           isPrimary: maps[i]['is_primary'] > 0 ? true : false);
     });
+  }
+
+  static Future<void> savePhotos(int recipeId, List<RecipePhoto> photos) async {
+    final Database db = await RecipeDatabase.instance.database;
+
+    Batch batch = db.batch();
+
+    photos.forEach((element) {
+      Map<String, dynamic> elements = element.toMap(recipeId);
+
+      batch.insert(RecipeDatabase.photosTable, elements,
+          conflictAlgorithm: ConflictAlgorithm.replace);
+    });
+
+    await batch.commit(noResult: true);
   }
 
   // static Future<RecipePhoto> getPrimaryImage(int recipeId) async {
@@ -47,7 +66,9 @@ class RecipePhotoDatabaseManager {
 
     List<int> ids = photos.map((e) => e.id).toList();
 
-    await db.delete(RecipeDatabase.photosTable,
-        where: "id IN (${ids.join(', ')})");
+    if (ids.isNotEmpty) {
+      await db.delete(RecipeDatabase.photosTable,
+          where: "id IN (${ids.join(', ')})");
+    }
   }
 }
